+++ MaketaTests/EdgesTests.swift
@@ -628,4 +628,44 @@ class EdgesTests: ConstraintsTestCase {
         XCTAssertEqual(bottom.firstAttribute, .bottomMargin)
         XCTAssertEqual(bottom.constant, -20)
     }
+	
+	// MARK: Autolayout not changed on right side
+	func testAutolayoutIsNotActivatedOnRightView() {
+		// given
+		let view = UIView()
+		superview.addSubview(view)
+		
+		let otherView = UIView()
+		superview.addSubview(otherView)
+		
+		XCTAssertTrue(view.translatesAutoresizingMaskIntoConstraints)
+		XCTAssertTrue(otherView.translatesAutoresizingMaskIntoConstraints)
+		
+		// when
+		view.mkt.edges = otherView.mkt.edges
+		
+		// then
+		XCTAssertFalse(view.translatesAutoresizingMaskIntoConstraints)
+		XCTAssertTrue(otherView.translatesAutoresizingMaskIntoConstraints)
+	}
+	
+	func testAutolayoutIsNotDeactivatedOnRightView() {
+		// given
+		let view = UIView()
+		superview.addSubview(view)
+		
+		let otherView = UIView()
+		otherView.preparedForAutolayout()
+		superview.addSubview(otherView)
+		
+		XCTAssertTrue(view.translatesAutoresizingMaskIntoConstraints)
+		XCTAssertFalse(otherView.translatesAutoresizingMaskIntoConstraints)
+		
+		// when
+		view.mkt.edges = otherView.mkt.edges
+		
+		// then
+		XCTAssertFalse(view.translatesAutoresizingMaskIntoConstraints)
+		XCTAssertFalse(otherView.translatesAutoresizingMaskIntoConstraints)
+	}
 }